+++ src/bible-search/RenderBibleDataFound.tsx
@@ -1,12 +1,12 @@
 import { copyToClipboard } from '../server/appHelper';
-import { toInputText } from '../server/bible-helpers/serverBibleHelpers2';
+import { toInputText } from '../helper/bible-helpers/serverBibleHelpers2';
 import { useState } from 'react';
-import { bookToKey } from '../server/bible-helpers/bibleInfoHelpers';
+import { bookToKey } from '../helper/bible-helpers/bibleInfoHelpers';
 import BibleItem from '../bible-list/BibleItem';
 import {
     ConsumeVerseType,
     consumeStartVerseEndVerse,
-} from '../helper/bibleHelpers';
+} from '../helper/bible-helpers/bibleHelpers';
 import { useAppEffect } from '../helper/debuggerHelpers';
 import { useStateSettingNumber } from '../helper/settingHelper';
 import RenderVersesOption from './RenderVersesOption';
